requirements.txt
static==1.0.2
gevent==0.13.7
gevent-socketio==0.3.5-rc2
socketIO-client==0.4
nose==1.3.0
pyScss==1.2.0
watchdog==0.7.1
